refactor: @ocap/shims -> @metamask/kernel-shims
