+++ packages/cli/package.json
@@ -36,11 +36,11 @@
     "@endo/bundle-source": "^4.0.0",
     "@endo/init": "^1.1.6",
     "@endo/promise-kit": "^1.1.10",
+    "@metamask/kernel-shims": "workspace:^",
     "@metamask/kernel-utils": "workspace:^",
     "@metamask/logger": "workspace:^",
     "@metamask/snaps-utils": "^9.1.0",
     "@metamask/utils": "^11.4.0",
-    "@ocap/shims": "workspace:^",
     "@types/node": "^22.13.1",
     "chokidar": "^4.0.1",
     "glob": "^11.0.0",
