+++ packages/cli/src/commands/watch.test.ts
@@ -1,4 +1,4 @@
-import '@ocap/shims/endoify';
+import '@metamask/kernel-shims/endoify';
 
 import { Logger } from '@metamask/logger';
 import { watch } from 'chokidar';
