+++ packages/extension/package.json
@@ -46,6 +46,7 @@
     "@metamask/json-rpc-engine": "^10.0.3",
     "@metamask/kernel-errors": "workspace:^",
     "@metamask/kernel-rpc-methods": "workspace:^",
+    "@metamask/kernel-shims": "workspace:^",
     "@metamask/kernel-store": "workspace:^",
     "@metamask/kernel-utils": "workspace:^",
     "@metamask/logger": "workspace:^",
@@ -55,7 +56,6 @@
     "@metamask/streams": "workspace:^",
     "@metamask/superstruct": "^3.2.1",
     "@metamask/utils": "^11.4.0",
-    "@ocap/shims": "workspace:^",
     "nanoid": "^5.1.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
