+++ packages/extension/tsconfig.build.json
@@ -16,7 +16,7 @@
     { "path": "../ocap-kernel/tsconfig.build.json" },
     { "path": "../logger/tsconfig.build.json" },
     { "path": "../kernel-rpc-methods/tsconfig.build.json" },
-    { "path": "../shims/tsconfig.build.json" },
+    { "path": "../kernel-shims/tsconfig.build.json" },
     { "path": "../kernel-store/tsconfig.build.json" },
     { "path": "../streams/tsconfig.build.json" },
     { "path": "../kernel-utils/tsconfig.build.json" }
