+++ packages/kernel-shims/README.md
@@ -1,0 +1,17 @@
+# `@metamask/kernel-shims`
+
+Ocap Kernel shims.
+
+Essentially, our particular flavor of SES `lockdown` etc.
+
+## Installation
+
+`yarn add @metamask/kernel-shims`
+
+or
+
+`npm install @metamask/kernel-shims`
+
+## Contributing
+
+This package is part of a monorepo. Instructions for contributing can be found in the [monorepo README](https://github.com/MetaMask/ocap-kernel#readme).
