+++ packages/kernel-shims/package.json
@@ -1,27 +1,37 @@
 {
-  "name": "@ocap/shims",
+  "name": "@metamask/kernel-shims",
   "version": "0.0.0",
-  "private": true,
-  "description": "Shims for ocap programming",
+  "description": "Ocap Kernel shims",
+  "keywords": [
+    "MetaMask",
+    "object capabilities",
+    "ocap"
+  ],
+  "homepage": "https://github.com/MetaMask/ocap-kernel/tree/main/packages/kernel-shims#readme",
+  "bugs": {
+    "url": "https://github.com/MetaMask/ocap-kernel/issues"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/MetaMask/ocap-kernel.git"
   },
+  "license": "SEE LICENSE IN LICENSE",
   "sideEffects": false,
   "type": "module",
   "exports": {
     "./endoify": "./dist/endoify.js",
     "./package.json": "./package.json"
   },
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.mjs",
+  "main": "./dist/endoify.js",
+  "module": "./dist/endoify.js",
   "files": [
     "dist/"
   ],
   "scripts": {
     "build": "node scripts/bundle.js",
     "build:docs": "typedoc",
-    "changelog:validate": "../../scripts/validate-changelog.sh @ocap/shims",
+    "changelog:update": "../../scripts/update-changelog.sh @metamask/kernel-shims",
+    "changelog:validate": "../../scripts/validate-changelog.sh @metamask/kernel-shims",
     "clean": "rimraf --glob './*.tsbuildinfo' ./.eslintcache ./coverage ./dist",
     "lint": "yarn lint:eslint && yarn lint:misc --check && yarn constraints && yarn lint:dependencies",
     "lint:dependencies": "depcheck",
@@ -70,5 +80,9 @@
   },
   "engines": {
     "node": "^20 || >=22"
+  },
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://registry.npmjs.org/"
   }
 }
