+++ packages/kernel-shims/src/endoify.js
