+++ packages/kernel-shims/typedoc.json
