+++ packages/nodejs/src/vat/vat-worker.ts
@@ -1,4 +1,4 @@
-import '@ocap/shims/endoify';
+import '@metamask/kernel-shims/endoify';
 
 import { Logger } from '@metamask/logger';
 import type { VatId } from '@metamask/ocap-kernel';
