+++ packages/nodejs/test/e2e/kernel-worker.test.ts
@@ -1,4 +1,4 @@
-import '@ocap/shims/endoify';
+import '@metamask/kernel-shims/endoify';
 
 import { Kernel } from '@metamask/ocap-kernel';
 import type { VatConfig, VatId } from '@metamask/ocap-kernel';
