--- packages/shims/README.md
@@ -1,7 +1,0 @@
-# `@ocap/shims`
-
-Shims for ocap programming that we want to import without transforming, such as `lockdown` and `eventual-send`.
-
-## Contributing
-
-This package is part of a monorepo. Instructions for contributing can be found in the [monorepo README](https://github.com/MetaMask/ocap-kernel#readme).
